Removed autofac dependant IClassInstanceManager implementation and moved the interface to the libs module. The assembly loader now scans for an implementation of this interface in the user's project (similar to IScreenGrabber).
