+++ Lib/Gauge.CSharp.Lib.csproj
@@ -56,6 +56,7 @@
     <Compile Include="DataStoreFactory.cs" />
     <Compile Include="DataStoreType.cs" />
     <Compile Include="GaugeMessages.cs" />
+    <Compile Include="IClassInstanceManager.cs" />
     <Compile Include="IScreenGrabber.cs" />
     <Compile Include="MessageCollector.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
