+++ Lib/IClassInstanceManager.cs
@@ -1,10 +1,13 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Reflection;
 
-
-namespace Gauge.CSharp.Runner.InstanceManagement
+namespace Gauge.CSharp.Lib
 {
     public interface IClassInstanceManager
     {
+        void Initialize(List<Assembly> assemblies);
+
 
         object Get(Type declaringType);
