+++ Runner/Gauge.CSharp.Runner.csproj
@@ -76,9 +76,6 @@
   <ItemGroup>
     <Compile Include="AssemblyLocater.cs" />
     <Compile Include="AssemblyLoader.cs" />
-    <Compile Include="InstanceManagement\ClassInstanceManagerFactory.cs" />
-    <Compile Include="InstanceManagement\ContainerBuilder\DiContainerBuilder.cs" />
-    <Compile Include="InstanceManagement\DependencyInjectingInstanceManager.cs" />
     <Compile Include="InstanceManagement\DefaultClassInstanceManager.cs" />
     <Compile Include="Converters\IParamConverter.cs" />
     <Compile Include="DefaultScreenGrabber.cs" />
@@ -102,8 +99,6 @@
     <Compile Include="IGaugeCommand.cs">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="InstanceManagement\IClassInstanceManager.cs" />
-    <Compile Include="InstanceManagement\ContainerBuilder\IDiContainerBuilder.cs" />
     <Compile Include="ISandbox.cs" />
     <Compile Include="IStepRegistry.cs" />
     <Compile Include="LogConfiguration.cs" />
