+++ Runner/IAssemblyLoader.cs
@@ -25,6 +25,7 @@ namespace Gauge.CSharp.Runner
     {
         List<Assembly> AssembliesReferencingGaugeLib { get; }
         List<Type> ScreengrabberTypes { get; }
+        List<Type> ClassInstanceManagerTypes { get; }
         Assembly GetTargetLibAssembly();
         List<MethodInfo> GetMethods(Type annotationType);
     }
