--- Runner/InstanceManagement/ContainerBuilder/IDiContainerBuilder.cs
@@ -1,9 +1,0 @@
-﻿using Autofac;
-
-namespace Gauge.CSharp.Runner.InstanceManagement.ContainerBuilder
-{
-    public interface IDiContainerBuilder
-    {
-        IContainer BuildContainer();
-    }
-}
