+++ Runner/packages.config
@@ -1,6 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Autofac" version="3.5.2" targetFramework="net45" />
   <package id="Gauge.CSharp.Core" version="0.0.3.0" targetFramework="net45" />
   <package id="Gauge.CSharp.Lib" version="0.5.2.0" targetFramework="net45" />
   <package id="Google.ProtocolBuffers" version="2.4.1.555" targetFramework="net45" />
